+++ saleor/webhook/event_types.py
@@ -732,6 +732,8 @@ class WebhookEventSyncType:
     TRANSACTION_INITIALIZE_SESSION = "transaction_initialize_session"
     TRANSACTION_PROCESS_SESSION = "transaction_process_session"
 
+    LIST_STORED_PAYMENT_METHODS = "list_stored_payment_methods"
+
     EVENT_MAP: dict[str, dict[str, Any]] = {
         PAYMENT_LIST_GATEWAYS: {
             "name": "List payment gateways",
@@ -803,6 +805,10 @@ class WebhookEventSyncType:
         },
         TRANSACTION_PROCESS_SESSION: {
             "name": "Process transaction session",
+            "permission": PaymentPermissions.HANDLE_PAYMENTS,
+        },
+        LIST_STORED_PAYMENT_METHODS: {
+            "name": "List tokenized payment methods that can be used by the customer.",
             "permission": PaymentPermissions.HANDLE_PAYMENTS,
         },
     }