+++ CHANGELOG.md
@@ -85,6 +85,7 @@ Shipping methods can be removed by the user after it has been assigned to a chec
   - `checkoutShippingAddressUpdate`, `checkoutBillingAddressUpdate`, `shopAddressUpdate`, `warehouseUpdate`
   - Add `metadata` to `AddressInput` field
 - Add support for unhandled field `Shop.trackInventoryByDefault` for `productVariantCreate` and `productVariantBulkCreate` - #13492 by @Smit-Parmar
+- Add `storedPaymentMethods` field to `User` and `Checkout` type - #13581
 - - Add `AttributeBulkUpdate` mutation - #13532 by @SzymJ
 
 - Better checkout error feedback - #13458 by @korycins